authenticatable moved to Utilities directory
