+++ src/edu/saddleback/cs4b/Backend/Utilitys/Authenticatable.java
@@ -1,4 +1,4 @@
-package edu.saddleback.cs4b.Backend.Server;
+package edu.saddleback.cs4b.Backend.Utilitys;
 
 public interface Authenticatable {
     String getKey();
